% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sfc_linestring_featureset_2d}
\alias{sfc_linestring_featureset_2d}
\alias{sfc_linestring_featureset_3d}
\alias{sfc_multilinestring_featureset_2d}
\alias{sfc_multilinestring_featureset_3d}
\alias{sfc_multipoint_featureset_2d}
\alias{sfc_multipoint_featureset_3d}
\alias{sfc_multipolygon_featureset_2d}
\alias{sfc_multipolygon_featureset_3d}
\alias{sfc_point_featureset_2d}
\alias{sfc_point_featureset_3d}
\alias{sfc_polygon_featureset_2d}
\alias{sfc_polygon_featureset_3d}
\alias{as_featureset_2d}
\alias{as_featureset_3d}
\title{Create a \code{FeatureSet} with 2-dimensional geometries}
\usage{
sfc_linestring_featureset_2d(x, sr)

sfc_linestring_featureset_3d(x, sr)

sfc_multilinestring_featureset_2d(x, sr)

sfc_multilinestring_featureset_3d(x, sr)

sfc_multipoint_featureset_2d(x, sr)

sfc_multipoint_featureset_3d(x, sr)

sfc_multipolygon_featureset_2d(x, sr)

sfc_multipolygon_featureset_3d(x, sr)

sfc_point_featureset_2d(x, sr)

sfc_point_featureset_3d(x, sr)

sfc_polygon_featureset_2d(x, sr)

sfc_polygon_featureset_3d(x, sr)

as_featureset_2d(attrs, geoms, n, sr)

as_featureset_3d(attrs, geoms, n, sr, has_z)
}
\arguments{
\item{sr}{a list representing the spatial reference e.g. \code{list(wkid = 3857)}}

\item{attrs}{a data.frame of columns that define the attributes}

\item{geoms}{an sfc geometry column}

\item{n}{the number of rows in \code{attrs}, if geometry is provided, must have as many elements as \code{n}}

\item{has_z}{bool. Whether Z or M geometries are provided. When \code{has_z = FALSE}, the attribute \code{hasM = true}}
}
\description{
Create a \code{FeatureSet} with 2-dimensional geometries
}
