% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sfg_point_as_point}
\alias{sfg_point_as_point}
\alias{sfg_multipoint_as_multipoint}
\alias{sfg_linestring_as_polyline}
\alias{sfg_multilinestring_as_polyline}
\alias{sfg_polygon_as_polygon}
\alias{sfg_multipolygon_as_polygon}
\title{Convert sfg geometries to EsriJSON strings}
\usage{
sfg_point_as_point(x)

sfg_multipoint_as_multipoint(x)

sfg_linestring_as_polyline(x)

sfg_multilinestring_as_polyline(x)

sfg_polygon_as_polygon(x)

sfg_multipolygon_as_polygon(x)
}
\arguments{
\item{x}{an sfg object}
}
\description{
Convert sfg geometries to EsriJSON strings
}
