% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sfc_linestring_features_2d}
\alias{sfc_linestring_features_2d}
\alias{sfc_linestring_features_3d}
\alias{sfc_multilinestring_features_2d}
\alias{sfc_multilinestring_features_3d}
\alias{sfc_multipoint_features_2d}
\alias{sfc_multipoint_features_3d}
\alias{sfc_multipolygon_features_2d}
\alias{sfc_multipolygon_features_3d}
\alias{sfc_point_features_2d}
\alias{sfc_point_features_3d}
\alias{sfc_polygon_features_2d}
\alias{sfc_polygon_features_3d}
\title{Create an EsriJSON feature array}
\usage{
sfc_linestring_features_2d(x)

sfc_linestring_features_3d(x)

sfc_multilinestring_features_2d(x)

sfc_multilinestring_features_3d(x)

sfc_multipoint_features_2d(x)

sfc_multipoint_features_3d(x)

sfc_multipolygon_features_2d(x)

sfc_multipolygon_features_3d(x)

sfc_point_features_2d(x)

sfc_point_features_3d(x)

sfc_polygon_features_2d(x)

sfc_polygon_features_3d(x)
}
\arguments{
\item{an}{sfc geometry vector}
}
\description{
Create an EsriJSON feature array
}
