% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sf_as_features_2d}
\alias{sf_as_features_2d}
\alias{sf_as_features_3d}
\alias{sfc_linestring_features_2d}
\alias{sfc_linestring_features_3d}
\alias{sfc_multilinestring_features_2d}
\alias{sfc_multilinestring_features_3d}
\alias{sfc_multipoint_features_2d}
\alias{sfc_multipoint_features_3d}
\alias{sfc_multipolygon_features_2d}
\alias{sfc_multipolygon_features_3d}
\alias{sfc_point_features_2d}
\alias{sfc_point_features_3d}
\alias{sfc_polygon_features_2d}
\alias{sfc_polygon_features_3d}
\title{Create an EsriJSON feature array}
\usage{
sf_as_features_2d(attrs, geoms, n, sr)

sf_as_features_3d(attrs, geoms, n, sr)

sfc_linestring_features_2d(x, sr)

sfc_linestring_features_3d(x, sr)

sfc_multilinestring_features_2d(x, sr)

sfc_multilinestring_features_3d(x, sr)

sfc_multipoint_features_2d(x, sr)

sfc_multipoint_features_3d(x, sr)

sfc_multipolygon_features_2d(x, sr)

sfc_multipolygon_features_3d(x, sr)

sfc_point_features_2d(x, sr)

sfc_point_features_3d(x, sr)

sfc_polygon_features_2d(x, sr)

sfc_polygon_features_3d(x, sr)
}
\arguments{
\item{attrs}{a data.frame of columns that define the attributes}

\item{geoms}{an sfc geometry column}

\item{n}{the number of rows in \code{attrs}, if geometry is provided, must have as many elements as \code{n}}

\item{sr}{a list representing the spatial reference e.g. \code{list(wkid = 3857)}}

\item{an}{sfc geometry vector}
}
\description{
Create an EsriJSON feature array
}
